src/RepositoryService/requirements.txt
flask
flask-cors
pycrypto
pycryptodome
pyjwt
gunicorn
flasgger
redis
mysql-connector-python